requirements.txt
requests
Pillow
beautifulsoup4
selenium